requirements.txt
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
WTForms>=3.1.2
Werkzeug>=3.1.0
SQLAlchemy>=2.0.36
pyodbc>=5.2.0
python-dotenv>=1.0.1
email-validator>=2.1.1
gunicorn>=22.0.0
